ticket-service/requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
pydantic
python-multipart
python-dotenv
